LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-style-changes-008-expected.txt

É
É
É

FAIL .before 1 assert_equals: 
<div id="item1" data-expected-width="20" data-expected-height="50" data-offset-y="0" class=" before" data-offset-x="5">É</div>
offsetLeft expected 5 but got 2
PASS .before 2 
PASS .before 3 
FAIL .after 4 assert_equals: 
<div id="item1" data-expected-width="20" data-expected-height="50" data-offset-y="0" class=" before after" data-offset-x="10">É</div>
offsetLeft expected 10 but got 4
PASS .after 5 
FAIL .after 6 assert_equals: 
<div id="item3" data-expected-width="30" data-expected-height="50" data-offset-y="100" class=" before after" data-offset-x="5">É</div>
offsetLeft expected 5 but got 2
